+++ app/build.gradle
@@ -15,6 +15,7 @@ android {
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 16
         versionName '2.6.0'
+        applicationId 'exe.bbll8.aeolus'
     }
 
     buildFeatures {
@@ -50,12 +51,12 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
 
             if (useAnemoConfig) {
-                signingConfig = signingConfigs.named('anemo').get()
+                signingConfig = signingConfigs.named('exe.bbllw8.aeolus').get()
             }
         }
         debug {
             if (useAnemoConfig) {
-                signingConfig = signingConfigs.named('anemo').get()
+                signingConfig = signingConfigs.named('exe.bbllw8.aeolus').get()
             }
         }
     }
