+++ app/src/main/java/exe/bbllw8/aeolus/config/password/PasswordDialog.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2021 2bllw8
  * SPDX-License-Identifier: GPL-3.0-only
  */
-package exe.bbllw8.anemo.config.password;
+package exe.bbllw8.aeolus.config.password;
 
 import android.app.Activity;
 import android.app.AlertDialog;
@@ -12,8 +12,8 @@ import android.graphics.drawable.Drawable;
 import androidx.annotation.LayoutRes;
 import androidx.annotation.StringRes;
 
-import exe.bbllw8.anemo.R;
-import exe.bbllw8.anemo.lock.LockStore;
+import exe.bbllw8.aeolus.R;
+import exe.bbllw8.aeolus.lock.LockStore;
 
 public abstract class PasswordDialog {
     protected static final int MIN_PASSWORD_LENGTH = 4;
