+++ app/src/main/java/exe/bbllw8/aeolus/documents/home/HomeEnvironment.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2021 2bllw8
  * SPDX-License-Identifier: GPL-3.0-only
  */
-package exe.bbllw8.anemo.documents.home;
+package exe.bbllw8.aeolus.documents.home;
 
 import android.content.Context;
 
@@ -11,9 +11,9 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 
 public final class HomeEnvironment {
-    public static final String AUTHORITY = "exe.bbllw8.anemo.documents";
+    public static final String AUTHORITY = "exe.bbllw8.aeolus.documents";
 
-    public static final String ROOT = "anemo";
+    public static final String ROOT = "aeolus";
     public static final String ROOT_DOC_ID = "root";
 
     private final Path baseDir;
