+++ app/src/main/java/exe/bbllw8/aeolus/lock/LockStore.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2021 2bllw8
  * SPDX-License-Identifier: GPL-3.0-only
  */
-package exe.bbllw8.anemo.lock;
+package exe.bbllw8.aeolus.lock;
 
 import android.app.job.JobInfo;
 import android.app.job.JobScheduler;
