+++ app/src/main/java/exe/bbllw8/aeolus/shell/LauncherActivity.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2022 2bllw8
  * SPDX-License-Identifier: GPL-3.0-only
  */
-package exe.bbllw8.anemo.shell;
+package exe.bbllw8.aeolus.shell;
 
 import android.app.Activity;
 import android.content.Intent;
@@ -17,10 +17,10 @@ import androidx.annotation.Nullable;
 import java.util.Arrays;
 import java.util.Optional;
 
-import exe.bbllw8.anemo.R;
-import exe.bbllw8.anemo.documents.home.HomeEnvironment;
-import exe.bbllw8.anemo.lock.LockStore;
-import exe.bbllw8.anemo.lock.UnlockActivity;
+import exe.bbllw8.aeolus.R;
+import exe.bbllw8.aeolus.documents.home.HomeEnvironment;
+import exe.bbllw8.aeolus.lock.LockStore;
+import exe.bbllw8.aeolus.lock.UnlockActivity;
 
 public class LauncherActivity extends Activity {
     // https://cs.android.com/android/platform/superproject/+/master:packages/apps/DocumentsUI/AndroidManifest.xml
